+++ sale_order_general_discount/models/res_partner.py
@@ -3,14 +3,10 @@
 
 from odoo import fields, models
 
-from odoo.addons import decimal_precision as dp
-
 
 class ResPartner(models.Model):
     _inherit = "res.partner"
 
     sale_discount = fields.Float(
-        digits=dp.get_precision("Discount"),
-        string="Discount (%)",
-        company_dependent=True,
+        string="Discount", digits="Discount", company_dependent=True,
     )
