+++ sale_order_general_discount/readme/CONTRIBUTORS.rst
@@ -1,2 +1,3 @@
 * Sergio Teruel <sergio.teruel@tecnativa.com>
 * Raf Ven <raf.ven@dynapps.be>
+* Sudhir Arya <sudhir@erpharbor.com>
